feat : 유저가 참여하고 있는 파티 디스플레이 기능 구현 세부사항 : 1. 기능 이름 명명 : shortcut 2. 기존 디렉토리에서 분리 후 새 디렉토리 생성 3. 몇몇 파일 이름 수정
